requirements.txt
alembic==1.11.2
aniso8601==9.0.1
anyio==3.7.1
backoff==2.2.1
boto3==1.28.22
botocore==1.31.22
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==14.0
confuse==2.0.1
croniter==1.4.1
dagit==1.4.4
dagster==1.4.4
dagster-graphql==1.4.4
dagster-postgres==0.20.4
dagster-webserver==1.4.4
docstring-parser==0.15
exceptiongroup==1.1.2
fsspec==2023.6.0
gql==3.4.1
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
grpcio==1.56.2
grpcio-health-checking==1.56.2
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mws==0.8.15
packaging==23.1
pendulum==2.1.2
protobuf==4.24.0
psycopg2-binary==2.9.7
pydantic==1.10.12
python-amazon-sp-api==0.18.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.31.0
tabulate==0.9.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.0
typing_extensions==4.7.1
universal-pathlib==0.0.24
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2